compile.sh添加错误判断

+++ compile.sh
@@ -7,5 +7,9 @@ if [ ! -f "a.asm" ]; then
 fi
 dd if=/dev/zero of=myhd.img bs=512 count=10240
 nasm -f bin a.asm -o a.bin
+if [ ! $? ]; then
+    echo "nasm出现错误"
+    exit 1
+fi
 dd if=a.bin of=myhd.img bs=512 count=1 conv=notrunc
 bochs -f mybochsrc
